requirements.txt
yt-dlp
assemblyai
moviepy
python-dotenv
requests
